+++ packages/atomic/src/components/atomic-refine-modal/atomic-refine-modal.tsx
@@ -18,6 +18,7 @@ import {
   BindStateToController,
   InitializableComponent,
   InitializeBindings,
+  DeferUntilRender,
 } from '../../utils/initialization-utils';
 import CloseIcon from 'coveo-styleguide/resources/icons/svg/close.svg';
 import {getFacetElements, SortDropdownOption} from '../../utils/store';
@@ -49,6 +50,7 @@ export class AtomicRefineModal implements InitializableComponent {
   private breadcrumbManager!: BreadcrumbManager;
   public querySummary!: QuerySummary;
   private facetManager!: FacetManager;
+  private closeButton?: HTMLElement;
   @InitializeBindings() public bindings!: Bindings;
   @Element() public host!: HTMLElement;
 
@@ -64,6 +66,8 @@ export class AtomicRefineModal implements InitializableComponent {
   @State() @BindStateToController('sort') public sortState!: SortState;
   @State() public error!: Error;
 
+  @Prop({mutable: true}) openButton?: HTMLElement;
+
   @Prop({reflect: true, mutable: true}) modalStatus!: ModalStatus;
   @Watch('modalStatus')
   watchEnabled(modalStatus: ModalStatus) {
@@ -73,12 +77,24 @@ export class AtomicRefineModal implements InitializableComponent {
       case 'opened':
         document.body.classList.add(modalOpenedClass);
         this.duplicateFacetElements();
+        this.focusOnCloseButton();
         break;
       case 'closed':
         document.body.classList.remove(modalOpenedClass);
         this.flushFacetElements();
+        this.focusOnOpenButton();
         break;
     }
+  }
+
+  @DeferUntilRender()
+  private focusOnCloseButton() {
+    this.closeButton?.focus();
+  }
+
+  @DeferUntilRender()
+  private focusOnOpenButton() {
+    this.openButton?.focus();
   }
 
   public initialize() {
@@ -127,6 +143,7 @@ export class AtomicRefineModal implements InitializableComponent {
           class="grid place-items-center"
           part="close-button"
           onClick={() => this.dismiss()}
+          ref={(closeButton) => (this.closeButton = closeButton)}
         >
           <atomic-icon class="w-5 h-5" icon={CloseIcon}></atomic-icon>
         </Button>
@@ -160,9 +177,9 @@ export class AtomicRefineModal implements InitializableComponent {
     }
 
     return [
-      <div part="section-title" class="text-2xl font-bold truncate mb-3 mt-8">
+      <h1 part="section-title" class="text-2xl font-bold truncate mb-3 mt-8">
         {this.bindings.i18n.t('sort')}
-      </div>,
+      </h1>,
       <div class="relative">
         <select
           class="btn-outline-neutral w-full cursor-pointer text-lg font-bold flex-grow appearance-none rounded-lg px-6 py-5"
@@ -186,9 +203,9 @@ export class AtomicRefineModal implements InitializableComponent {
 
     return [
       <div class="w-full flex justify-between mt-8 mb-3">
-        <span part="section-title" class="text-2xl font-bold truncate">
+        <h1 part="section-title" class="text-2xl font-bold truncate">
           {this.bindings.i18n.t('filters')}
-        </span>
+        </h1>
         {this.breadcrumbManagerState.hasBreadcrumbs && (
           <Button
             onClick={() => this.breadcrumbManager.deselectAll()}
@@ -237,27 +254,32 @@ export class AtomicRefineModal implements InitializableComponent {
       return;
     }
 
+    const isOpened = this.modalStatus === 'opened';
+
     return (
-      <div
-        part="container"
-        class={`w-screen h-screen fixed flex flex-col justify-between bg-background text-on-background left-0 top-0 z-10 ${
-          this.modalStatus === 'opened'
-            ? 'animate-scaleUpRefineModal'
-            : 'animate-scaleDownRefineModal'
-        }`}
-      >
-        <div class="px-6">{this.renderHeader()}</div>
-        <hr class="border-neutral"></hr>
-        <div class="overflow-auto px-6 flex-grow">
-          <div class="adjust-for-scroll-bar">
-            <div class="centered">
-              {this.renderSort()}
-              {this.renderFilters()}
+      <atomic-focus-trap active={isOpened}>
+        <div
+          part="container"
+          class={`w-screen h-screen fixed flex flex-col justify-between bg-background text-on-background left-0 top-0 z-10 ${
+            isOpened
+              ? 'animate-scaleUpRefineModal'
+              : 'animate-scaleDownRefineModal'
+          }`}
+          aria-modal={isOpened.toString()}
+        >
+          <div class="px-6">{this.renderHeader()}</div>
+          <hr class="border-neutral"></hr>
+          <div class="overflow-auto px-6 flex-grow">
+            <div class="adjust-for-scroll-bar">
+              <div class="centered">
+                {this.renderSort()}
+                {this.renderFilters()}
+              </div>
             </div>
           </div>
+          {this.renderFooter()}
         </div>
-        {this.renderFooter()}
-      </div>
+      </atomic-focus-trap>
     );
   }
 }