+++ packages/headless/src/features/facets/facet-search-set/category/category-facet-search-request-builder.test.ts
@@ -23,84 +23,85 @@ describe('#buildCategoryFacetSearchRequest', () => {
 
   beforeEach(() => setupState());
 
-  it('retrieves the #captions from the categoryFacetSearchSet', () => {
+  it('retrieves the #captions from the categoryFacetSearchSet', async () => {
     const captions = {a: 'A'};
     state.categoryFacetSearchSet[id].options.captions = captions;
 
-    expect(buildParams().captions).toEqual(captions);
+    expect((await buildParams()).captions).toEqual(captions);
   });
 
-  it('retrieves the #numberOfValues from the categoryFacetSearchSet', () => {
+  it('retrieves the #numberOfValues from the categoryFacetSearchSet', async () => {
     const numberOfValues = 5;
     state.categoryFacetSearchSet[id].options.numberOfValues = numberOfValues;
 
-    expect(buildParams().numberOfValues).toEqual(numberOfValues);
+    expect((await buildParams()).numberOfValues).toEqual(numberOfValues);
   });
 
-  it('retrieves the #query from the categoryFacetSearchSet', () => {
+  it('retrieves the #query from the categoryFacetSearchSet', async () => {
     const query = 'hello';
     state.categoryFacetSearchSet[id].options.query = query;
 
-    expect(buildParams().query).toEqual(`*${query}*`);
+    expect((await buildParams()).query).toEqual(`*${query}*`);
   });
 
-  it('retrieves the #basePath fron the categoryFacetSet', () => {
+  it('retrieves the #basePath fron the categoryFacetSet', async () => {
     const basePath = ['a'];
     const request = buildMockCategoryFacetRequest({basePath});
     state.categoryFacetSet[id] = buildMockCategoryFacetSlice({request});
 
-    expect(buildParams().basePath).toBe(basePath);
+    expect((await buildParams()).basePath).toBe(basePath);
   });
 
-  it('retrieves the #field from the categoryFacetSet', () => {
+  it('retrieves the #field from the categoryFacetSet', async () => {
     const field = 'author';
     const request = buildMockCategoryFacetRequest({field});
     state.categoryFacetSet[id] = buildMockCategoryFacetSlice({request});
 
-    expect(buildParams().field).toBe(field);
+    expect((await buildParams()).field).toBe(field);
   });
 
-  it('retrieves #filterFacetCount from the categoryFacetSet', () => {
+  it('retrieves #filterFacetCount from the categoryFacetSet', async () => {
     const request = buildMockCategoryFacetRequest({filterFacetCount: true});
     state.categoryFacetSet[id] = buildMockCategoryFacetSlice({request});
 
-    expect(buildParams().filterFacetCount).toBe(true);
+    expect((await buildParams()).filterFacetCount).toBe(true);
   });
 
-  it('retrieves the #delimitingCharacter from the categoryFacetSet', () => {
+  it('retrieves the #delimitingCharacter from the categoryFacetSet', async () => {
     const delimitingCharacter = '|';
     const request = buildMockCategoryFacetRequest({delimitingCharacter});
     state.categoryFacetSet[id] = buildMockCategoryFacetSlice({request});
 
-    expect(buildParams().delimitingCharacter).toBe(delimitingCharacter);
+    expect((await buildParams()).delimitingCharacter).toBe(delimitingCharacter);
   });
 
-  it('sets the #searchContext to the search request params', () => {
+  it('sets the #searchContext to the search request params', async () => {
     const facet = state.categoryFacetSet[id]!.request;
-    const request = {...buildSearchRequest(state).request, facets: [facet]};
+    const builtRequest = await buildSearchRequest(state);
+    const request = {...builtRequest.request, facets: [facet]};
 
-    expect(buildParams().searchContext).toEqual({
+    expect((await buildParams()).searchContext).toEqual({
       ...request,
       visitorId: expect.any(String),
     });
   });
 
-  it('#ignorePaths is empty when currentValues is empty', () => {
+  it('#ignorePaths is empty when currentValues is empty', async () => {
     state.categoryFacetSet[id]!.request.currentValues = [];
-    expect(buildParams().ignorePaths).toEqual([]);
+    expect((await buildParams()).ignorePaths).toEqual([]);
   });
 
-  it('#ignorePaths returns the correct path when currentValues has one level', () => {
+  it('#ignorePaths returns the correct path when currentValues has one level', async () => {
     state.categoryFacetSet[id]!.request.currentValues = [
       buildMockCategoryFacetValueRequest({
         value: 'level1',
         state: 'selected',
       }),
     ];
-    expect(buildParams().ignorePaths).toEqual([['level1']]);
+    expect((await buildParams()).ignorePaths).toEqual([['level1']]);
   });
 
-  it('#ignorePaths returns the correct path when currentValues has more than one level', () => {
+  it('#ignorePaths returns the correct path when currentValues has more than one level', async () => {
     state.categoryFacetSet[id]!.request.currentValues = [
       buildMockCategoryFacetValueRequest({
         value: 'level1',
@@ -112,6 +113,6 @@ describe('#buildCategoryFacetSearchRequest', () => {
         ],
       }),
     ];
-    expect(buildParams().ignorePaths).toEqual([['level1', 'level2']]);
+    expect((await buildParams()).ignorePaths).toEqual([['level1', 'level2']]);
   });
 });