+++ packages/headless/src/features/search/search-request.test.ts
@@ -18,88 +18,88 @@ describe('search request', () => {
     state = createMockState();
   });
 
-  it('#searchRequest returns the state #query', () => {
+  it('#searchRequest returns the state #query', async () => {
     state.query.q = 'hello';
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.q).toBe(state.query.q);
   });
 
-  it('#searchRequest returns the state #enableQuerySyntax', () => {
+  it('#searchRequest returns the state #enableQuerySyntax', async () => {
     state.query.enableQuerySyntax = true;
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.enableQuerySyntax).toBe(state.query.enableQuerySyntax);
   });
 
-  it('#searchRequest returns the state #sortCriteria', () => {
+  it('#searchRequest returns the state #sortCriteria', async () => {
     state.sortCriteria = 'qre';
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.sortCriteria).toBe(state.sortCriteria);
   });
 
-  it('#searchRequest returns the state #numberOfResults', () => {
+  it('#searchRequest returns the state #numberOfResults', async () => {
     state.pagination.numberOfResults = 10;
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.numberOfResults).toBe(state.pagination.numberOfResults);
   });
 
-  it('#searchRequest returns the state #firstResult', () => {
+  it('#searchRequest returns the state #firstResult', async () => {
     state.pagination.firstResult = 10;
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.firstResult).toBe(state.pagination.firstResult);
   });
 
-  it('#searchRequest returns the state #dictionaryFieldContext.contextValues', () => {
+  it('#searchRequest returns the state #dictionaryFieldContext.contextValues', async () => {
     const contextValues = {price: 'cad'};
     state.dictionaryFieldContext.contextValues = contextValues;
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
 
     expect(params.dictionaryFieldContext).toBe(contextValues);
   });
 
-  it('#searchRequest returns the facets in the state #facetSet', () => {
+  it('#searchRequest returns the facets in the state #facetSet', async () => {
     const request = buildMockFacetRequest({field: 'objecttype'});
     state.facetSet[1] = request;
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toContainEqual(request);
   });
 
-  it('#searchRequest returns the facets in the state #numericFacetSet', () => {
+  it('#searchRequest returns the facets in the state #numericFacetSet', async () => {
     const request = buildMockNumericFacetRequest({field: 'objecttype'});
     state.numericFacetSet[1] = request;
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toContainEqual(request);
   });
 
-  it('#searchRequest returns the facets in the state #dateFacetSet', () => {
+  it('#searchRequest returns the facets in the state #dateFacetSet', async () => {
     const request = buildMockDateFacetRequest({field: 'objecttype'});
     state.dateFacetSet[1] = request;
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toContainEqual(request);
   });
 
-  it('#searchRequestParams returns the facets in the #categoryFacetSet', () => {
+  it('#searchRequestParams returns the facets in the #categoryFacetSet', async () => {
     const request = buildMockCategoryFacetRequest({field: 'objecttype'});
     state.categoryFacetSet[1] = buildMockCategoryFacetSlice({request});
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toContainEqual(request);
   });
 
-  it('when no facets are configured, the #searchRequestParams does not contain a #facets key', () => {
-    const request = buildSearchRequest(state).request;
+  it('when no facets are configured, the #searchRequestParams does not contain a #facets key', async () => {
+    const request = (await buildSearchRequest(state)).request;
     expect(request.facets).toBe(undefined);
   });
 
   it(`when there are facets ids in the same order as the facetOrder array,
-  #searchRequestParams orders the facets in the same order as the response`, () => {
+  #searchRequestParams orders the facets in the same order as the response`, async () => {
     const facetId1 = '1';
     const facetId2 = '2';
 
@@ -108,7 +108,7 @@ describe('search request', () => {
     state.facetSet[facetId1] = buildMockFacetRequest({facetId: facetId1});
     state.facetSet[facetId2] = buildMockFacetRequest({facetId: facetId2});
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toEqual([
       state.facetSet[facetId2],
       state.facetSet[facetId1],
@@ -116,7 +116,7 @@ describe('search request', () => {
   });
 
   it(`when there is a facet id that is not in the facetOrder array,
-  #searchRequestParams includes it at the end of the facets array`, () => {
+  #searchRequestParams includes it at the end of the facets array`, async () => {
     const facetId1 = '1';
     const facetId2 = '2';
 
@@ -125,96 +125,102 @@ describe('search request', () => {
     state.facetSet[facetId1] = buildMockFacetRequest({facetId: facetId1});
     state.facetSet[facetId2] = buildMockFacetRequest({facetId: facetId2});
 
-    const {facets} = buildSearchRequest(state).request;
+    const {facets} = (await buildSearchRequest(state)).request;
     expect(facets).toEqual([
       state.facetSet[facetId2],
       state.facetSet[facetId1],
     ]);
   });
 
-  it('#searchRequestParams returns the facetOptions in state', () => {
+  it('#searchRequestParams returns the facetOptions in state', async () => {
     state.facetOptions = buildMockFacetOptions({freezeFacetOrder: true});
 
-    const params = buildSearchRequest(state).request;
+    const params = (await buildSearchRequest(state)).request;
     expect(params.facetOptions).toEqual(state.facetOptions);
   });
 
-  it('should send visitorId if analytics is enable', () => {
+  it('should send visitorId if analytics is enable', async () => {
     state.configuration.analytics.enabled = true;
-    expect(buildSearchRequest(state).request.visitorId).toBeDefined();
+    expect((await buildSearchRequest(state)).request.visitorId).toBeDefined();
   });
 
-  it('should not send visitorId if analytics is disabled', () => {
+  it('should not send visitorId if analytics is disabled', async () => {
     state.configuration.analytics.enabled = false;
-    expect(buildSearchRequest(state).request.visitorId).not.toBeDefined();
+    expect(
+      (await buildSearchRequest(state)).request.visitorId
+    ).not.toBeDefined();
   });
 
-  it('#searchRequest.tab holds the #originLevel2', () => {
+  it('#searchRequest.tab holds the #originLevel2', async () => {
     const originLevel2 = 'youtube';
     state.configuration.analytics.originLevel2 = originLevel2;
-    expect(buildSearchRequest(state).request.tab).toBe(originLevel2);
+    expect((await buildSearchRequest(state)).request.tab).toBe(originLevel2);
   });
 
-  it('#searchRequest.referrer holds the #originLevel3', () => {
+  it('#searchRequest.referrer holds the #originLevel3', async () => {
     const originLevel3 = 'www.coveo.com';
     state.configuration.analytics.originLevel3 = originLevel3;
-    expect(buildSearchRequest(state).request.referrer).toBe(originLevel3);
+    expect((await buildSearchRequest(state)).request.referrer).toBe(
+      originLevel3
+    );
   });
 
-  it('#searchRequest.fieldsToInclude holds the #fieldsToInclude', () => {
+  it('#searchRequest.fieldsToInclude holds the #fieldsToInclude', async () => {
     state.fields.fieldsToInclude = ['foo', 'bar'];
-    expect(buildSearchRequest(state).request.fieldsToInclude).toEqual(
+    expect((await buildSearchRequest(state)).request.fieldsToInclude).toEqual(
       expect.arrayContaining(['foo', 'bar'])
     );
   });
 
-  it('#searchRequest.fieldsToInclude does not holds #fieldsToInclude if #fetchAllFields is active', () => {
+  it('#searchRequest.fieldsToInclude does not holds #fieldsToInclude if #fetchAllFields is active', async () => {
     state.fields.fieldsToInclude = ['foo', 'bar'];
     state.fields.fetchAllFields = true;
-    expect(buildSearchRequest(state).request.fieldsToInclude).not.toBeDefined();
+    expect(
+      (await buildSearchRequest(state)).request.fieldsToInclude
+    ).not.toBeDefined();
   });
 
-  it('when there are no cq expressions in state, cq is undefined', () => {
-    expect(buildSearchRequest(state).request.cq).toBe(undefined);
+  it('when there are no cq expressions in state, cq is undefined', async () => {
+    expect((await buildSearchRequest(state)).request.cq).toBe(undefined);
   });
 
-  it('when there is a cq expression, it sets the cq to the expression', () => {
+  it('when there is a cq expression, it sets the cq to the expression', async () => {
     state.advancedSearchQueries.cq = 'a';
-    expect(buildSearchRequest(state).request.cq).toBe('a');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a');
   });
 
-  it('when there is an active tab, it sets cq to the active tab expression', () => {
+  it('when there is an active tab, it sets cq to the active tab expression', async () => {
     state.tabSet.a = buildMockTabSlice({expression: 'a', isActive: true});
-    expect(buildSearchRequest(state).request.cq).toBe('a');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a');
   });
 
   it(`when there is cq and an active tab,
-  it sets the cq to a concatenated expression`, () => {
+  it sets the cq to a concatenated expression`, async () => {
     state.advancedSearchQueries.cq = 'a';
     state.tabSet.b = buildMockTabSlice({expression: 'b', isActive: true});
 
-    expect(buildSearchRequest(state).request.cq).toBe('a AND b');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a AND b');
   });
 
   it(`when the cq and active tab expressions are surrounded by spaces,
-  it trims them before concatenating`, () => {
+  it trims them before concatenating`, async () => {
     state.advancedSearchQueries.cq = ' a ';
     state.tabSet.b = buildMockTabSlice({expression: ' b ', isActive: true});
 
-    expect(buildSearchRequest(state).request.cq).toBe('a AND b');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a AND b');
   });
 
-  it('static filter with an active value, it sets cq to the active filter value expression', () => {
+  it('static filter with an active value, it sets cq to the active filter value expression', async () => {
     const value = buildMockStaticFilterValue({
       expression: 'a',
       state: 'selected',
     });
     state.staticFilterSet.a = buildMockStaticFilterSlice({values: [value]});
-    expect(buildSearchRequest(state).request.cq).toBe('a');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a');
   });
 
   it(`static filter with two active values,
-  it concatenates the expressions with OR and wraps them with parentheses`, () => {
+  it concatenates the expressions with OR and wraps them with parentheses`, async () => {
     const valueA = buildMockStaticFilterValue({
       expression: 'a',
       state: 'selected',
@@ -227,11 +233,11 @@ describe('search request', () => {
     state.staticFilterSet.a = buildMockStaticFilterSlice({
       values: [valueA, valueB],
     });
-    expect(buildSearchRequest(state).request.cq).toBe('(a OR b)');
+    expect((await buildSearchRequest(state)).request.cq).toBe('(a OR b)');
   });
 
   it(`static filter with two active values, one value has an empty space as an expression,
-  it filters off the empty expression`, () => {
+  it filters off the empty expression`, async () => {
     const valueA = buildMockStaticFilterValue({
       expression: 'a',
       state: 'selected',
@@ -244,6 +250,6 @@ describe('search request', () => {
     state.staticFilterSet.a = buildMockStaticFilterSlice({
       values: [valueA, valueB],
     });
-    expect(buildSearchRequest(state).request.cq).toBe('a');
+    expect((await buildSearchRequest(state)).request.cq).toBe('a');
   });
 });