+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.32.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.31.7...@coveo/headless@1.32.0) (2021-11-22)
+
+
+### Features
+
+* **atomic:** expose headless as a subpackage ([#1500](https://github.com/coveo/ui-kit/issues/1500)) ([4ac4b05](https://github.com/coveo/ui-kit/commit/4ac4b05f62ecd78fa08739cecb185fc611e2a2d6))
+
+
+
+
+
+## [1.31.7](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.31.6...@coveo/headless@1.31.7) (2021-11-18)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+## [1.31.6](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.31.5...@coveo/headless@1.31.6) (2021-11-15)
+
+
+### Bug Fixes
+
+* **headless:** fix visitorID parameter retrieved from coveo.analytics ([#1478](https://github.com/coveo/ui-kit/issues/1478)) ([c6adb69](https://github.com/coveo/ui-kit/commit/c6adb69310c785b4e60dc6f7eb6a6f9fcfba2a42))
+
+
+
+
+
+## [1.31.5](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.31.4...@coveo/headless@1.31.5) (2021-11-12)
+
+
+### Bug Fixes
+
+* **headless:** load sourcemaps from CDN ([#1462](https://github.com/coveo/ui-kit/issues/1462)) ([34c386e](https://github.com/coveo/ui-kit/commit/34c386edf23176db1e25842db394ed34fb748d35))
+* **headless:** send filterFacetCount param with facet search request ([#1475](https://github.com/coveo/ui-kit/issues/1475)) ([f83e143](https://github.com/coveo/ui-kit/commit/f83e143c44a455abaefd25d78c9413cee494afcd))
+
+
+
+
+
 ## [1.31.4](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.31.3...@coveo/headless@1.31.4) (2021-11-10)