+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.43.5](https://github.com/coveo/ui-kit/compare/headless-react@0.43.4...headless-react@0.43.5) (2021-11-22)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.43.4](https://github.com/coveo/ui-kit/compare/headless-react@0.43.3...headless-react@0.43.4) (2021-11-18)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.43.3](https://github.com/coveo/ui-kit/compare/headless-react@0.43.2...headless-react@0.43.3) (2021-11-15)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.43.2](https://github.com/coveo/ui-kit/compare/headless-react@0.43.1...headless-react@0.43.2) (2021-11-12)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.43.1](https://github.com/coveo/ui-kit/compare/headless-react@0.43.0...headless-react@0.43.1) (2021-11-10)
 
 **Note:** Version bump only for package headless-react